fixed an infinite

+++ includes/ft_printf_bonus.h
@@ -6,7 +6,7 @@
 /*   By: gakarbou <marvin@42.fr>                    +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2024/11/07 14:42:46 by gakarbou          #+#    #+#             */
-/*   Updated: 2024/11/09 22:47:32 by gakarbou         ###   ########.fr       */
+/*   Updated: 2024/11/10 00:02:52 by gakarbou         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -31,7 +31,7 @@ void			ft_putnbr_times(int long nb, int len);
 void			ft_putaddr_times(unsigned long nb, int len, char *base);
 
 int				ft_write_and_return(char c, int ret);
-int				ft_write_integer(int nb, int *flags, size_t res);
+int				ft_write_integer(int nb, int *flags, int res);
 int				ft_write_unsigned(unsigned int nb, int *flags);
 int				ft_write_addr(unsigned long addr, int *flags);
 int				ft_write_char(int c, int *flags);
