+++ srcs/writers_bonus.c
@@ -6,13 +6,13 @@
 /*   By: gakarbou <marvin@42.fr>                    +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2024/11/09 20:55:22 by gakarbou          #+#    #+#             */
-/*   Updated: 2024/11/09 21:56:03 by gakarbou         ###   ########.fr       */
+/*   Updated: 2024/11/10 00:04:42 by gakarbou         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
 #include "ft_printf_bonus.h"
 
-void	ft_handle_int_flags(int *flags, size_t *len, int nb)
+void	ft_handle_int_flags(int *flags, int *len, int nb)
 {
 	if (flags[6])
 		*len = ft_max(*len, flags[8]);
@@ -24,16 +24,16 @@ void	ft_handle_int_flags(int *flags, size_t *len, int nb)
 		flags[2] = 0;
 }
 
-int	ft_write_integer(int nb, int *flags, size_t res)
+int	ft_write_integer(int nb, int *flags, int res)
 {
-	size_t	len;
+	int	len;
 
 	len = ft_log(nb, 10);
 	ft_handle_int_flags(flags, &len, nb);
 	if (nb >= 0 && flags[2])
 		res += ft_write_and_return(' ', 1);
 	while (!flags[4] && flags[7] && !flags[5] && (len + res
-			+ (nb < 0 || flags[1] || (nb >= 0 && flags[2]))) < (size_t)flags[7])
+			+ (nb < 0 || flags[1] || (nb >= 0 && flags[2]))) < flags[7])
 		res += ft_write_and_return(' ', 1);
 	if (nb < 0)
 		res += ft_write_and_return('-', 1);
@@ -42,15 +42,15 @@ int	ft_write_integer(int nb, int *flags, size_t res)
 	if (flags[5])
 		len = ft_max(flags[7] - res, len);
 	ft_putnbr_times(nb, len);
-	while (flags[4] && ((len + res) < (size_t)flags[7]))
+	while (flags[4] && ((len + res) < flags[7]))
 		res += ft_write_and_return(' ', 1);
 	return (res + len);
 }
 
 int	ft_write_unsigned(unsigned int nb, int *flags)
 {
-	size_t	len;
-	size_t	res;
+	int	len;
+	int	res;
 
 	res = 0;
 	len = ft_log(nb, 10);
@@ -61,21 +61,21 @@ int	ft_write_unsigned(unsigned int nb, int *flags)
 	if ((flags[6] || flags[4]) && flags[5])
 		flags[5] = 0;
 	while (!flags[4] && flags[7] && !flags[5]
-		&& ((len + res) < (size_t)flags[7]))
+		&& ((len + res) < flags[7]))
 		res += ft_write_and_return(' ', 1);
 	if (flags[5])
 		len = ft_max(flags[7] - res, len);
 	ft_putnbr_times(nb, len);
-	while (flags[4] && ((len + res) < (size_t)flags[7]))
+	while (flags[4] && ((len + res) < flags[7]))
 		res += ft_write_and_return(' ', 1);
 	return (res + len);
 }
 
 int	ft_write_string(char *str, int *flags)
 {
-	size_t	res;
-	size_t	len;
-	size_t	i;
+	int	res;
+	int	len;
+	int	i;
 
 	if (!str)
 	{
